backend/requirements.txt
psycopg2-binary
pandas
duckdb
python-dotenv
fastapi
uvicorn[standard]
pydantic